# swagger.json
swagger: "2.0"
info:
  description: "This is an Express server used to manage identity provider services. Users manage services by creating requests handled by administrators. You can find out more about the project on the [github page](https://github.com/rciam/rciam-federation-registry) or you can try it out live  [here](http://service-registry.aai-dev.grnet.gr/egi). "
  version: "1.2.1"
  title: "Federation Registry"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Rciam"
    url: "https://github.com/rciam"
host: "swagger.io"
basePath: "/v2"
...
schemes:
- "https"
paths:
  /tenants/{name}/tags:
    get:
      tags:
      - Service Tags
      summary: "Get all tags"
      description: "Get all available tags for a given tenant"
      parameters:
        - in: query
          name: tag
          type: string
          description: "Search string to filter tags result"
          required: false
        - in: path
          name: name
          type: string
          description: "Tenant that the request is targeting"
          required: true
      responses: 
        "200":
          description: "OK"
          schema: 
            type: "array"
            items:
              type: "string"
        "401": 
          description: "Unauthenticated request" 
        "500":
          description: "Error"
  /tenants/{name}/tags/services/{id}:
    get:
      tags:
      - Service Tags
      summary: "Get all tags of a service"
      description: "Get all available tags attached to a service"
      parameters:
        - in: path
          name: id
          type: integer
          description: "Id of the requested service"
          required: true
        - in: path
          name: name
          type: string
          description: "Tenant that the request is targeting"
          required: true
      responses: 
        "200":
          description: "OK"
          schema: 
            type: "array"
            items:
              type: "string"
        "401": 
          description: "Unauthenticated request" 
        "500":
          description: "Error"
    post:
      tags:
      - Service Tags
      summary: "Create new tags for a service"
      description: "Create new tags for a service"
      parameters:
        - in: path
          name: id
          type: integer
          description: "Id of the requested service"
          required: true
        - in: path
          name: name
          type: string
          description: "Tenant that the request is targeting"
          required: true
        - in: body
          name: body
          schema:
            type: "array"
            description: "Array of string tags"
            items: 
              type: "string"
      responses: 
        "200":
          description: "OK"
        "401": 
          description: "Unauthenticated request" 
        "500":
          description: "Error"
    delete:
      tags:
      - Service Tags
      summary: "Delete serivce tags"
      description: "Delete tags for a service"
      parameters:
        - in: path
          name: id
          type: integer
          description: "Id of the requested service"
          required: true
        - in: path
          name: name
          type: string
          description: "Tenant that the request is targeting"
          required: true
        - in: body
          name: body
          schema:
            type: "array"
            description: "Array of string tags"
            items: 
              type: "string"
      responses: 
        "200":
          description: "OK"
        "401": 
          description: "Unauthenticated request" 
        "500":
          description: "Error"
...
  /tenants/{name}/services/{id}:
    get:
      tags:
      - "Services"
      summary: "Get service by id"
      description: "Gets service by id"
      operationId: "getService"
      produces:
      - "application/json"
      parameters:
      - in: path
        type: string
        name: name
        description: "Tenant that the request is targeting"
        required: true
      - name: "id"
        in: "path"
        description: "ID of service to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful"
          schema:
            $ref: "#/definitions/Service"
        "401":
          description: "Unauthenticated request"
        "500":
          description: "Internal Error"
      security:
      - access_tokenAuth: []
...
  /tenants/{name}/services:
    get:
      tags:
      - "Services"
      summary: "Get all available services"
      description: "Returns all available services, some properties are returned only for authorized requests"
      operationId: "getServices"
      produces:
      - "application/json"
      parameters:
        - in: path
          type: string
          name: name
          description: "Tenant that the request is targeting"
          required: true
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              type: "object"
              $ref: "#/definitions/Service"
        "401":
          description: "Unauthenticated request"
        "500":
          description: "Internal Error"
      security:
      - access_tokenAuth: []
    post:
      tags:
      - "Services"
      summary: "Add services"
      description: "Endpoint used to directly add services to federation registry. Should be used only when necessary and not as an alternative way to add services."
      operationId: "addServices"
      parameters:
        - in: path
          type: string
          name: name
          description: "Tenant that the request is targeting"
          required: true
        - in: "body"
          name: "body"
          description: "Service details to add to the registry"
          required: true
          schema:
            $ref: "#/definitions/PostServices"
      responses:
        "200":
          description: "OK"
        "422":
          description: "Unprocessable Entry"
        "401":
          description: "Unauthenticated request"
        "500":
          description: "Internal Error"
...
securityDefinitions:
  access_tokenAuth:
    type: "basic"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
  admin_key:
    type: "apiKey"
    name: "X-API-KEY"
    in: "header"
definitions:
...
  PostServices:
    type: "array"
    items:
      type: "object"
      required:
        - integration_environment
        - protocol
        - client_id
        - metadata_url
      $ref: "#/definitions/Service"
...
  Service:
    type: "object"
    properties:
      service_name:
        type: "string"
        description: "Human-readable application name"
      service_description:
        type: "string"
        description: "Human-readable text description"
      logo_uri:
        type: "string"
        format: "uri"
        description: "URL that points to an image"
      policy_uri:
        type: "string"
        format: "uri"
        description: "URL for the Policy Statement of this service"
      aup_uri:
        type: "string"
        format: "uri"
        description: "URL that contains the Acceptable Use Policy for the users of the service"
      organization_id:
        type: "integer"
        description: "The unique identifier of the Organization of the Service"
      integration_environment:
        type: "string"
        enum:
        - "demo"
        - "development"
        - "production"
      contacts:
        type: "array"
        items:
          type: "object"
          properties:
            email:
              type: "string"
              format: "email"
            type:
              type: "string"
              enum:
              - "admin"
              - "support"
              - "technical"
              - "security"
      protocol:
        type: "string"
        enum:
        - "oidc"
        - "saml"
      entity_id:
        type: "string"
        description: "Unique identifier for saml services"
      metadata_uri:
        type: "string"
        format: "uri"
        description: "Metadata uri for saml services"
      client_id:
        type: "string"
        description: "Unique identifier for oidc services"
      grant_types:
        type: "array"
        items:
          type: "string"
          enum:
          - "authorization_code"
          - "client_credentials"
          - "urn:ietf:params:oauth:grant-type:token-exchange"
          - "urn:ietf:params:oauth:grant-type:device_code"
          - "implicit"
        description: "Grant Types Values"
      scope:
        type: "array"
        items:
          type: "string"
          enum:
          - "openid"
          - "email"
          - "profile"
          - "offline_access"
          - "eduperson_entitlement"
          - "eduperson_scoped_affiliation"
          - "eduperson_unique_id"
          - "refeds_edu"
        description: "oidc specific"
      allow_introspection:
        type: "boolean"
        description: "Allow calls to the Introspection Endpoint. For oidc services."
      code_challenge_method:
        type: "string"
        enum:
        - "plain"
        - "S256"
        - "null"
        description: "Proof Key for Code Exchange (PKCE) Code Challenge Method. For oidc services."
      client_secret:
        type: "string"
        format: "password"
        description: "Client secret, oidc specific"
      token_endpoint_auth_method:
        type: "string"
        enum:
        - "client_secret_basic"
        - "client_secret_post"
        - "client_secret_jwt"
        - "private_key_jwt"
        - "none"
        description: "Authentication method for token endpoint, oidc specific"
      token_endpoint_auth_signing_alg:
        type: "string"
        enum:
          - "HS256"
          - "HS384"
          - "HS512"
          - "RS256"
          - "RS384"
          - "RS512"
          - "PS256"
        description: "Signing algorithm for the token endpoint, oidc specific"
      jwks:
        type: "object"
        properties:
          keys:
            type: "array"
            items:
              type: "string"
        description: "Public key value for the token endpoint, oidc specific"
      jwks_uri:
        type: "string"
        description: "Uri containing public key for the token endpoint, oidc specific"
      country:
        type: "string"
        description: "Country of the organisation of the service"
      website_url:
        type: "string"
        description: "Website Url containing information about the service"
      redirect_uris:
        type: "array"
        items:
          type: "string"
          description: "Allowed Redirect uris for the service"
      generate_client_secret:
        type: boolean
        description: "Prompts backend to generate value for client secret. Oidc specific"
      reuse_refresh_tokens:
        type: "boolean"
        description: "Adds the offline_access scope to the client's scope, oidc specific"
      clear_access_tokens_on_refresh:
        type: "boolean"
        description: "oidc specific"
      device_code_validity_seconds:
        type: integer
        minimum: 0
        maximum: 1800
        description: "oidc specific"
      access_token_validity_seconds:
        type: integer
        minimum: 0
        maximum: 11160
        description: "oidc specific"
      refresh_token_validity_seconds:
        type: integer
        minimum: 0
        maximum: 1422000
        description: "oidc specific"
      id_token_timeout_seconds:
        type: integer
        minimum: 0
        maximum: 3600
        description: "oidc specific"
      code_of_condact_field:
        type: "boolean"
        description: "There could be multiple code of contact fields with different depending on the tenant some of them might be required to be true."
...